fix(protocol): replace unsafe `as u32` integer casts with checked conversions and fix SSRF bypass
Replace all bare `as u32` truncating casts with `u32::try_from(x).unwrap_or(u32::MAX)` across
the protocol stack to prevent silent truncation on 64-bit platforms. This hardens the HKDF info
construction (session key derivation, gossip summaries, HRW hashing, sender subkey derivation,
holder proof serialization), FCPS frame encoding, symbol store counts, and E2E assertion summaries.

Because the HKDF info buffer now consistently length-prefixes all variable-length fields with
u32::try_from (matching the canonical wire format), the session test vectors' expected derived
keys are updated accordingly.

Additional fixes in this commit:
- fcp-sandbox/egress: use to_ipv4_mapped() instead of to_ipv4() to prevent SSRF bypasses via
  IPv4-compatible IPv6 addresses (::x.x.x.x) which are distinct from IPv4-mapped (::ffff:x.x.x.x)
- fcp-sandbox/wasi: update ResourceLimiter::table_growing signature from u32 to usize to match
  the wasmtime 29+ API change
- fcp-host: add 'static bound to ConnectorRegistry generic in doctor_handler to satisfy
  axum's Send + 'static requirements
- fcp-protocol/symbol_envelope: change derive_sender_subkey parameter from &ZoneKey to &AeadKey
  to match actual usage at call sites
- fcp-store/repair: add #[allow(clippy::cast_possible_truncation)] annotation for safe u8→u32 cast

+++ crates/fcp-sandbox/src/wasi.rs
@@ -605,9 +605,9 @@ impl wasmtime::ResourceLimiter for WasiHostState {
 
     fn table_growing(
         &mut self,
-        _current: u32,
-        _desired: u32,
-        _maximum: Option<u32>,
+        _current: usize,
+        _desired: usize,
+        _maximum: Option<usize>,
     ) -> Result<bool, anyhow::Error> {
         // Tables are allowed to grow per default bounds
         Ok(true)
